backend/requirements.txt
Django>=4.2,<4.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary>=2.9
gunicorn>=20.1.0
python-dotenv>=1.0.0
django-cors-headers>=4.3 